Move Extent to separate file
